[FIX] payment/saas~16.5: avoid cte
It looks like a CTE limitation that cannot guarantee the order of the
expressions is respected.

> Trying to update the same row twice in a single statement is not
supported. Only one of the modifications takes place, but it is not easy
(and sometimes not possible) to reliably predict which one. This also
applies to deleting a row that was already updated in the same
statement: only the update is performed. Therefore you should generally
avoid trying to modify a single row twice in a single statement. In
particular avoid writing WITH sub-statements that could affect the same
rows changed by the main statement or a sibling sub-statement. The
effects of such a statement will not be predictable.

https://www.postgresql.org/docs/current/queries-with.html

Practical run below, first the steps withing a transaction to keep the
data, after the actual query ran by the script.
```
1282262_afu_17.0=> begin;                                                                                                                                                                                  [11/17611]
             SELECT o.id
               FROM payment_provider p
               JOIN payment_provider o
                 ON o.code = p.code
                AND o.module_id = p.module_id
                AND o.id != p.id
              WHERE p.id = 14
                AND true
 ;
             DELETE FROM payment_method_payment_provider_rel r
                   WHERE r.payment_provider_id = ANY(ARRAY[18])
 ;
         SELECT r.payment_method_id, UNNEST(ARRAY[18])
           FROM payment_method_payment_provider_rel r
          WHERE r.payment_provider_id = 14;
 ;
         INSERT INTO payment_method_payment_provider_rel (payment_method_id, payment_provider_id)
         SELECT r.payment_method_id, UNNEST(ARRAY[18])
           FROM payment_method_payment_provider_rel r
          WHERE r.payment_provider_id = 14;

 rollback;
         WITH duplicated_providers AS (
             SELECT o.id
               FROM payment_provider p
               JOIN payment_provider o
                 ON o.code = p.code
                AND o.module_id = p.module_id
                AND o.id != p.id
              WHERE p.id = 14
                AND true
         ),
         _cleanup AS (
             DELETE FROM payment_method_payment_provider_rel r
                   USING duplicated_providers d
                   WHERE r.payment_provider_id = d.id
         )
         INSERT INTO payment_method_payment_provider_rel (payment_method_id, payment_provider_id)
         SELECT r.payment_method_id, d.id
           FROM payment_method_payment_provider_rel r,
                duplicated_providers d
          WHERE r.payment_provider_id = 14
BEGIN
+----+
| id |
|----|
| 18 |
+----+
SELECT 1
DELETE 3
+-------------------+--------+
| payment_method_id | unnest |
|-------------------+--------|
| 1                 | 18     |
| 2                 | 18     |
+-------------------+--------+
SELECT 2
INSERT 0 2
ROLLBACK
duplicate key value violates unique constraint "payment_icon_payment_provider_rel_pkey"
DETAIL:  Key (payment_provider_id, payment_method_id)=(18, 1) already exists.
```

From the logs
```
2024-02-01 18:26:32,525 28 INFO 1282262_afu_17.0 odoo.modules.migration: module payment: Running migration [saas~16.5.2.0>] post-migrate
2024-02-01 18:26:33,442 28 ERROR 1282262_afu_17.0 odoo.sql_db: bad query:
        WITH duplicated_providers AS (
            SELECT o.id
              FROM payment_provider p
              JOIN payment_provider o
                ON o.code = p.code
               AND o.module_id = p.module_id
               AND o.id != p.id
             WHERE p.id = 14
               AND true
        ),
        _cleanup AS (
            DELETE FROM payment_method_payment_provider_rel r
                  USING duplicated_providers d
                  WHERE r.payment_provider_id = d.id
        )
        INSERT INTO payment_method_payment_provider_rel (payment_method_id, payment_provider_id)
        SELECT r.payment_method_id, d.id
          FROM payment_method_payment_provider_rel r,
               duplicated_providers d
         WHERE r.payment_provider_id = 14

ERROR: duplicate key value violates unique constraint "payment_icon_payment_provider_rel_pkey"
DETAIL:  Key (payment_provider_id, payment_method_id)=(18, 6) already exists.
```

tbg-1053
upg-1217754
upg-1282262
upg-1282954

closes odoo/upgrade#5626

+++ migrations/payment/saas~16.5.2.0/post-migrate.py
@@ -116,27 +116,32 @@ def copy_payment_methods_to_duplicated_providers(cr, xmlid, *, custom_mode=None)
         custom_mode_filter = "o.custom_mode = %(custom_mode)s"
         args["custom_mode"] = custom_mode
 
-    query = f"""
-        WITH duplicated_providers AS (
-            SELECT o.id
-              FROM payment_provider p
-              JOIN payment_provider o
-                ON o.code = p.code
-               AND o.module_id = p.module_id
-               AND o.id != p.id
-             WHERE p.id = %(provider)s
-               AND {custom_mode_filter}
-        ),
-        _cleanup AS (
-            DELETE FROM payment_method_payment_provider_rel r
-                  USING duplicated_providers d
-                  WHERE r.payment_provider_id = d.id
-        )
-        INSERT INTO payment_method_payment_provider_rel (payment_method_id, payment_provider_id)
-        SELECT r.payment_method_id, d.id
-          FROM payment_method_payment_provider_rel r,
-               duplicated_providers d
-         WHERE r.payment_provider_id = %(provider)s
-    """
+    cr.execute(
+        f"""
+        SELECT ARRAY_AGG(o.id)
+          FROM payment_provider p
+          JOIN payment_provider o
+            ON o.code = p.code
+           AND o.module_id = p.module_id
+           AND o.id != p.id
+         WHERE p.id = %(provider)s
+           AND {custom_mode_filter}
+        """,
+        args,
+    )
+    dup_ids = cr.fetchone()[0]
+    if not dup_ids:
+        return
+    args["ids"] = dup_ids
+    cr.execute(
+        """
+        DELETE FROM payment_method_payment_provider_rel r
+              WHERE r.payment_provider_id = ANY(%(ids)s);
 
-    cr.execute(query, args)
+        INSERT INTO payment_method_payment_provider_rel (payment_method_id, payment_provider_id)
+        SELECT r.payment_method_id, UNNEST(%(ids)s)
+          FROM payment_method_payment_provider_rel r
+         WHERE r.payment_provider_id = %(provider)s
+        """,
+        args,
+    )
